Close PreparedStatements and ResultSets (MUDetect finding 18 & 19).
The PreparedStatements originalPost and replies and the ResultSets postResult and repliesResult are left open. This might lead to a resource leak, since the connection is managed outside the method and might be left open for a long time, such that many open statements/result sets accumulate. See https://stackoverflow.com/questions/4507440/must-jdbc-resultsets-and-statements-be-closed-separately-although-the-connection