Close ResultSet in error case (MUDetectXP finding 12 & 13).
The ResultSet privResult is left open, if an exception occurs while fetching results in lines 291 or 305. This might lead to a resource leak with some JDBC drivers or to a “maximum open sursors exceeded” exception, in case too many result sets are left unclosed. See https://stackoverflow.com/questions/4507440/must-jdbc-resultsets-and-statements-be-closed-separately-although-the-connection